rename search endpoint to 'GET /variations' and move variations stuff into its own router. Also remove CNV-related prose in endpoint text
